requirements.txt
PyYAML==6.0.2
Telethon==1.36.0
